% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled-sigmoid.R
\name{dt_activation_scaled_sigmoid}
\alias{dt_activation_scaled_sigmoid}
\title{Scaled sigmoid activation}
\usage{
dt_activation_scaled_sigmoid(object, name = NULL, trainable = TRUE)
}
\arguments{
\item{object}{Model or layer object}

\item{name}{An optional name string for the layer. Should be unique in a model
(do not reuse the same name twice). It will be autogenerated if it isn't provided.}

\item{trainable}{Whether the layer weights will be updated during training.}
}
\description{
Scaled sigmoid activation
}
