% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-evaluation.R
\name{dt_compute_metrics}
\alias{dt_compute_metrics}
\title{Compute performance metrics}
\usage{
dt_compute_metrics(predictions)
}
\arguments{
\item{predictions}{Predictions data frame.}
}
\description{
Compute performance metrics
}
