% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{dt_compute_predictions}
\alias{dt_compute_predictions}
\title{Computes predictions}
\usage{
dt_compute_predictions(model, data)
}
\arguments{
\item{model}{A trained DeepTriangle model.}

\item{data}{A data frame, i.e. output of \code{dt_data_prep()}.}
}
\description{
Computes predictions
}
