% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early-stopping.R
\name{dt_early_stopping}
\alias{dt_early_stopping}
\title{Custom early stopping callback}
\usage{
dt_early_stopping(window_width = 10, patience = 100)
}
\arguments{
\item{window_width}{Window width for moving average calcuation, defaults to \code{10}.}

\item{patience}{Patience, defaults to \code{100}.}
}
\description{
Implements the early stopping callback detailed in the paper.
}
