% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_group_codes}
\alias{dt_group_codes}
\title{Group codes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 200 rows and 2 columns.}
\usage{
dt_group_codes
}
\description{
Data frame of group codes taken from Meyers (2015).
}
\keyword{datasets}
