% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-engineering.R
\name{dt_mutate_time_series}
\alias{dt_mutate_time_series}
\title{Add time series feature to data frame}
\usage{
dt_mutate_time_series(x)
}
\arguments{
\item{x}{A data frame.}
}
\description{
Add time series feature to data frame
}
