% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-epochs.R
\name{dt_optimize_epochs}
\alias{dt_optimize_epochs}
\title{Optimize number of epochs to train}
\usage{
dt_optimize_epochs(model, training_data, validation_data, batch_size = 128)
}
\arguments{
\item{model}{Keras model object, i.e. output of \code{dt_model()}.}

\item{training_data}{Training dataset.}

\item{validation_data}{Validation dataset.}

\item{batch_size}{Batch size.}
}
\description{
Optimize number of epochs to train
}
