% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dt_plot_predictions}
\alias{dt_plot_predictions}
\title{Plot predictions}
\usage{
dt_plot_predictions(predictions, group_code, lob, type = c("paid_loss",
  "claims_outstanding"))
}
\arguments{
\item{predictions}{Predictions data frame. Output of \code{dt_compute_predictions()}.}

\item{group_code}{Company code to plot.}

\item{lob}{LOB to plot}

\item{type}{One of \code{"paid_loss"} and \code{"claims_outstanding"}.}
}
\description{
Plot predictions
}
