% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-evaluation.R
\name{dt_tabulate_metrics}
\alias{dt_tabulate_metrics}
\title{Tabulate performance metrics for all models}
\usage{
dt_tabulate_metrics(data, metric = c("mape", "rmspe"))
}
\arguments{
\item{data}{Model results in tidy format.}

\item{metric}{Performance metric.}
}
\description{
Tabulate performance metrics for all models
}
