% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{dt_train_validation_split}
\alias{dt_train_validation_split}
\title{Convert data frame into train-validation lists for training}
\usage{
dt_train_validation_split(data)
}
\arguments{
\item{data}{Data frame corresponding to a single LOB, typically the output of \code{dt_data_prep()}.}
}
\description{
Convert data frame into train-validation lists for training
}
