% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stochastic_model_results}
\alias{stochastic_model_results}
\title{Stochastic model results}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 4 columns.}
\usage{
stochastic_model_results
}
\description{
Data frame of model results taken from Meyers (2015)
}
\keyword{datasets}
